Moves CSV source and destination test case to integration test file.
